Enhance /update
Include capacity management, make fields optional
