+++ src/cli/register-slash.ts
@@ -2,7 +2,6 @@
 import { REST } from "@discordjs/rest";
 import { APIUser, Routes } from "discord-api-types/v10";
 import dotenv from "dotenv";
-import { CapacityCommand } from "../slash/capacity";
 import { ChannelCommand } from "../slash/channel";
 import { CreateCommand } from "../slash/create";
 import { DeckCommand } from "../slash/deck";
@@ -23,7 +22,6 @@ export const classes = [
 	InfoCommand,
 	ForceDropCommand,
 	DropCommand,
-	CapacityCommand,
 	DeckCommand,
 	FinishCommand
 ];
