+++ src/slash/update.ts
@@ -19,11 +19,10 @@ export class UpdateCommand extends AutocompletableCommand {
 			.setDMPermission(false)
 			.setDefaultMemberPermissions(0)
 			.addStringOption(tournamentOption)
-			.addStringOption(option =>
-				option.setName("name").setDescription("The name of the tournament.").setRequired(true)
-			)
-			.addStringOption(option =>
-				option.setName("desc").setDescription("A description of the tournament.").setRequired(true)
+			.addStringOption(option => option.setName("name").setDescription("The name of the tournament."))
+			.addStringOption(option => option.setName("desc").setDescription("A description of the tournament."))
+			.addNumberOption(option =>
+				option.setName("capacity").setDescription("The new capacity for the tournament.").setMinValue(0)
 			)
 			.toJSON();
 	}
@@ -44,14 +43,47 @@ export class UpdateCommand extends AutocompletableCommand {
 			// rejection messages handled in helper
 			return;
 		}
-		const name = interaction.options.getString("name", true);
-		const description = interaction.options.getString("description", true);
 
-		tournament.name = name;
-		tournament.description = description;
+		const name = interaction.options.getString("name");
+		const description = interaction.options.getString("description");
+		const rawCap = interaction.options.getNumber("capacity");
+
+		let updated = false;
+
+		if (name) {
+			tournament.name = name;
+			updated = true;
+		}
+		if (description) {
+			tournament.description = description;
+			updated = true;
+		}
+
+		if (rawCap) {
+			// enforce integer cap
+			const capacity = Math.floor(rawCap);
+			const playerCount = tournament.participants.filter(p => p.deck?.approved).length;
+			// cap 0 means uncapped
+			if (capacity > 0 && playerCount > capacity) {
+				await interaction.reply(
+					`You have more players (${playerCount}) registered than the new cap. Please drop enough players then try again.`
+				);
+			}
+
+			tournament.participantLimit = capacity;
+			updated = true;
+		}
+		if (!updated) {
+			await interaction.reply(`You must provide at least one detail to update a tournament.`);
+			return;
+		}
 
 		await tournament.save();
 
-		await interaction.reply(`Tournament ${name} updated with following description!\n${description}`);
+		await interaction.reply(
+			`Tournament updated with the following details:\nName: ${tournament.name}\nDescription: ${
+				tournament.description
+			}\nCapacity: ${tournament.participantLimit === 0 ? "Uncapped" : tournament.participantLimit}`
+		);
 	}
 }
